삭제 기능 및 서버 예외처리 기능 추가

+++ src/main/java/DataView/project/service/PostingService.java
@@ -1,14 +1,13 @@
 package DataView.project.service;
 
+import DataView.project.domain.Member;
 import DataView.project.domain.Posting;
-import DataView.project.dto.CustomUserDetails;
 import DataView.project.repository.SDJpaPostingRepository;
 import jakarta.transaction.Transactional;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
 
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Optional;
 
 
 @Transactional
@@ -19,12 +18,9 @@ public class PostingService {
         this.postingRepository = postingRepository;
     }
 
-    public void postSave(Posting posting) {
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        CustomUserDetails userDetails = (CustomUserDetails) authentication.getPrincipal();
-
-        posting.setUserId(userDetails.getUsername());
-        posting.setAuthor(userDetails.getName());
+    public void postSave(Member member, Posting posting) {
+        posting.setUserId(member.getUsername());
+        posting.setAuthor(member.getName());
         posting.setCreated_at(LocalDateTime.now());
 
         postingRepository.save(posting);
@@ -34,7 +30,17 @@ public class PostingService {
         return postingRepository.findAllByBoardType(boardType);
     }
 
-    public void delete(Long postingId) {
-        postingRepository.deleteById(postingId);
+    public Posting getPostingById(Long id) {
+        Optional<Posting> posting = postingRepository.findById(id);
+        return posting.orElse(null);
+    }
+
+    public boolean deletePosting(Member member, Long postingId) {
+        if (getPostingById(postingId).getUserId().equals(member.getUsername())) {
+            postingRepository.deleteById(postingId);
+            return true;
+        } else {
+            return false;
+        }
     }
 }
